+++ tests/unit/data/tScience.m
@@ -11,12 +11,12 @@ classdef tScience < matlab.unittest.TestCase
         function magnitude(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            expectedMagnitude = sqrt(sum([data.X.^2, data.Y.^2, data.Z.^2], 2));
+            expectedMagnitude = sqrt(sum([science.X.^2, science.Y.^2, science.Z.^2], 2));
 
             % Exercise.
-            actualMagnitude = data.B;
+            actualMagnitude = science.B;
 
             % Verify.
             testCase.verifyThat(actualMagnitude, matlab.unittest.constraints.IsEqualTo(expectedMagnitude, Within = matlab.unittest.constraints.AbsoluteTolerance(1e-12)), ...
@@ -27,10 +27,10 @@ classdef tScience < matlab.unittest.TestCase
         function derivative_empty(testCase, DerivativeVariable)
 
             % Set up.
-            data = testCase.createEmptyTestData();
+            science = testCase.createEmptyTestData();
 
             % Exercise.
-            derivative = data.(DerivativeVariable);
+            derivative = science.(DerivativeVariable);
 
             % Verify.
             testCase.verifyEmpty(derivative, "Derivative of empty value should itself be empty.");
@@ -40,13 +40,13 @@ classdef tScience < matlab.unittest.TestCase
         function derivative_nonEmpty(testCase, DerivativeVariable)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
             v = erase(DerivativeVariable, "d");
-            expectedDerivative = diff(data.(v));
+            expectedDerivative = diff(science.(v));
 
             % Exercise.
-            actualDerivative = data.(DerivativeVariable);
+            actualDerivative = science.(DerivativeVariable);
 
             % Verify.
             testCase.verifyEqual(actualDerivative(1:end-1), expectedDerivative, "Derivative should match expected value.");
@@ -57,114 +57,114 @@ classdef tScience < matlab.unittest.TestCase
         function cropMethod_duration(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            expectedTimes = data.Time(2:end);
-            expectedData = data.DependentVariables(2:end, :);
+            expectedTimes = science.Time(2:end);
+            expectedData = science.DependentVariables(2:end, :);
 
             timeFilter = minutes(1);
 
             % Exercise and verify.
-            testCase.cropAndVerify(data, timeFilter, expectedTimes, expectedData);
+            testCase.cropAndVerify(science, timeFilter, expectedTimes, expectedData);
         end
 
         % Test that "crop" method crops data based on a "timerange" object.
         function cropMethod_timerange(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            expectedTimes = data.Time(3:end);
-            expectedData = data.DependentVariables(3:end, :);
+            expectedTimes = science.Time(3:end);
+            expectedData = science.DependentVariables(3:end, :);
 
-            timeFilter = timerange(data.Time(2), data.Time(end), "openleft");
+            timeFilter = timerange(science.Time(2), science.Time(end), "openleft");
 
             % Exercise and verify.
-            testCase.cropAndVerify(data, timeFilter, expectedTimes, expectedData);
+            testCase.cropAndVerify(science, timeFilter, expectedTimes, expectedData);
         end
 
         % Test that "crop" method crops data based on a "withtol" object.
         function cropMethod_withtol(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            expectedTimes = data.Time(4:6);
-            expectedData = data.DependentVariables(4:6, :);
+            expectedTimes = science.Time(4:6);
+            expectedData = science.DependentVariables(4:6, :);
 
-            timeFilter = withtol(data.Time(5), minutes(1));
+            timeFilter = withtol(science.Time(5), minutes(1));
 
             % Exercise and verify.
-            testCase.cropAndVerify(data, timeFilter, expectedTimes, expectedData);
+            testCase.cropAndVerify(science, timeFilter, expectedTimes, expectedData);
         end
 
         % Test that "crop" method also crops events.
         function cropMethod_events(testCase)
 
             % Set up.
-            data = testCase.createTestData();
-            data.Data.Properties.Events = eventtable(data.Data);
+            science = testCase.createTestData();
+            science.Data.Properties.Events = eventtable(science.Data);
 
             % Exercise.
-            data.crop(minutes(1));
+            science.crop(minutes(1));
 
             % Verify.
-            testCase.assertEqual(height(data.Events.Time), height(data.Time), "Data should be cropped as expected.");
-            testCase.verifyEqual(data.Events.Time, data.Time, "Data should be cropped as expected.");
+            testCase.assertEqual(height(science.Events.Time), height(science.Time), "Data should be cropped as expected.");
+            testCase.verifyEqual(science.Events.Time, science.Time, "Data should be cropped as expected.");
         end
 
         % Test that "crop" method does not fail when no data is selected.
         function cropMethod_noSelection(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
             % Exercise.
-            data.crop(timerange(datetime("Inf", TimeZone = "local"), datetime("-Inf", TimeZone = "local")));
+            science.crop(timerange(datetime("Inf", TimeZone = "local"), datetime("-Inf", TimeZone = "local")));
 
             % Verify.
-            testCase.verifyEmpty(data.IndependentVariable, "All data should be cropped out.");
-            testCase.verifyEmpty(data.DependentVariables, "All data should be cropped out.");
+            testCase.verifyEmpty(science.IndependentVariable, "All data should be cropped out.");
+            testCase.verifyEmpty(science.DependentVariables, "All data should be cropped out.");
 
-            testCase.verifyTrue(ismissing(data.MetaData.Timestamp), "All data should be cropped out.");
+            testCase.verifyTrue(ismissing(science.MetaData.Timestamp), "All data should be cropped out.");
         end
 
         % Test that "resample" method can resample to a lower frequency.
         function resampleMethod_lowerFrequency(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            initialFrequency = 1 / seconds(mode(data.dT));
+            initialFrequency = 1 / seconds(mode(science.dT));
 
             % Exercise.
-            resampledData = data.copy();
+            resampledData = science.copy();
             resampledData.resample(initialFrequency / 2);
 
             % Verify.
-            testCase.assertEqual(height(resampledData.IndependentVariable), height(data.Time) / 2, "Frequency should be halved.");
-            testCase.verifyEqual(resampledData.Time, (data.Time(1):minutes(2):data.Time(end))', "Frequency should be halved.");
+            testCase.assertEqual(height(resampledData.IndependentVariable), height(science.Time) / 2, "Frequency should be halved.");
+            testCase.verifyEqual(resampledData.Time, (science.Time(1):minutes(2):science.Time(end))', "Frequency should be halved.");
 
-            testCase.assertEqual(height(resampledData.DependentVariables), height(data.DependentVariables) / 2, "Frequency should be halved.");
+            testCase.assertEqual(height(resampledData.DependentVariables), height(science.DependentVariables) / 2, "Frequency should be halved.");
         end
 
         % Test that "resample" method can resample to a higher frequency.
         function resampleMethod_higherFrequency(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            initialFrequency = 1 / seconds(mode(data.dT));
+            initialFrequency = 1 / seconds(mode(science.dT));
 
             % Exercise.
-            resampledData = data.copy();
+            resampledData = science.copy();
             resampledData.resample(2 * initialFrequency);
 
             % Verify.
-            testCase.assertEqual(height(resampledData.IndependentVariable), (2 * height(data.Time)) - 1, "Frequency should be doubled.");
-            testCase.verifyEqual(resampledData.Time, (data.Time(1):seconds(30):data.Time(end))', "Frequency should be doubled.");
+            testCase.assertEqual(height(resampledData.IndependentVariable), (2 * height(science.Time)) - 1, "Frequency should be doubled.");
+            testCase.verifyEqual(resampledData.Time, (science.Time(1):seconds(30):science.Time(end))', "Frequency should be doubled.");
 
-            testCase.assertEqual(height(resampledData.DependentVariables), (2 * height(data.DependentVariables)) - 1, "Frequency should be doubled.");
+            testCase.assertEqual(height(resampledData.DependentVariables), (2 * height(science.DependentVariables)) - 1, "Frequency should be doubled.");
         end
 
         % Test that "resample" method throws when target frequency is not
@@ -172,29 +172,29 @@ classdef tScience < matlab.unittest.TestCase
         function resampleMethod_error(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
             % Exercise and verify.
-            testCase.verifyError(@() data.resample(0.12345), "", "Resampling should error when frequencies do not match.");
+            testCase.verifyError(@() science.resample(0.12345), "", "Resampling should error when frequencies do not match.");
         end
 
         % Test that "downsample" method can downsample to a lower frequency.
         function downsampleMethod_lowerFrequency(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            initialFrequency = 1 / seconds(mode(data.dT));
+            initialFrequency = 1 / seconds(mode(science.dT));
 
             % Exercise.
-            downsampledData = data.copy();
+            downsampledData = science.copy();
             downsampledData.downsample(initialFrequency / 2);
 
             % Verify.
-            testCase.assertEqual(height(downsampledData.IndependentVariable), height(data.Time) / 2, "Frequency should be halved.");
-            testCase.verifyEqual(downsampledData.Time, (data.Time(1):minutes(2):data.Time(end))', "Frequency should be halved.");
+            testCase.assertEqual(height(downsampledData.IndependentVariable), height(science.Time) / 2, "Frequency should be halved.");
+            testCase.verifyEqual(downsampledData.Time, (science.Time(1):minutes(2):science.Time(end))', "Frequency should be halved.");
 
-            testCase.assertEqual(height(downsampledData.DependentVariables), height(data.DependentVariables) / 2, "Frequency should be halved.");
+            testCase.assertEqual(height(downsampledData.DependentVariables), height(science.DependentVariables) / 2, "Frequency should be halved.");
             testCase.verifyTrue(any(ismissing(downsampledData.XYZ), "all"), "Initial data should be replaced with missing values to account for filter warm-up.");
         end
 
@@ -203,10 +203,10 @@ classdef tScience < matlab.unittest.TestCase
         function downsampleMethod_error(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
             % Exercise and verify.
-            testCase.verifyError(@() data.downsample(0.12345), "", "Resampling should error when frequencies do not match.");
+            testCase.verifyError(@() science.downsample(0.12345), "", "Resampling should error when frequencies do not match.");
         end
 
         % Test that "filter" method can filter with a "digitalFilter"
@@ -214,40 +214,40 @@ classdef tScience < matlab.unittest.TestCase
         function filterMethod_digitalFilter(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            initialFrequency = 1 / seconds(mode(data.dT));
+            initialFrequency = 1 / seconds(mode(science.dT));
             filter = designfilt("lowpassfir", SampleRate = initialFrequency, PassbandFrequency = (initialFrequency / 4), StopbandFrequency = 1.5 * (initialFrequency / 4));
 
             % Exercise.
-            downsampledData = data.copy();
-            downsampledData.filter(filter);
+            downsampledScience = science.copy();
+            downsampledScience.filter(filter);
 
             % Verify.
-            testCase.assertSize(downsampledData.IndependentVariable, size(data.Time), "Filtering should not affect size.");
-            testCase.verifyEqual(downsampledData.Time, data.Time, "Filtering should not affect time.");
+            testCase.assertSize(downsampledScience.IndependentVariable, size(science.Time), "Filtering should not affect size.");
+            testCase.verifyEqual(downsampledScience.Time, science.Time, "Filtering should not affect time.");
 
-            testCase.assertSize(downsampledData.DependentVariables, size(data.DependentVariables), "Filtering should not affect size.");
-            testCase.verifyTrue(all(ismissing(downsampledData.XYZ), "all"), "All data should be replaced with missing values to account for filter warm-up.");
+            testCase.assertSize(downsampledScience.DependentVariables, size(science.DependentVariables), "Filtering should not affect size.");
+            testCase.verifyTrue(all(ismissing(downsampledScience.XYZ), "all"), "All data should be replaced with missing values to account for filter warm-up.");
         end
 
         % Test that "replace" method replaces data with default filler.
         function replaceMethod_default(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
             % Exercise.
-            modifiedData = data.copy();
-            modifiedData.replace(minutes(3));
+            modifiedScience = science.copy();
+            modifiedScience.replace(minutes(3));
 
             % Verify.
-            testCase.assertSize(modifiedData.IndependentVariable, size(data.Time), "Time should not be modified.");
-            testCase.verifyEqual(modifiedData.Time, data.Time, "Time should not be modified.");
+            testCase.assertSize(modifiedScience.IndependentVariable, size(science.Time), "Time should not be modified.");
+            testCase.verifyEqual(modifiedScience.Time, science.Time, "Time should not be modified.");
 
-            testCase.assertSize(modifiedData.DependentVariables, size(data.DependentVariables), "Data size should not change.");
-            testCase.verifyTrue(all(ismissing(modifiedData.XYZ(1:4, :)), "all"), "Data within filter should be replaced.");
-            testCase.verifyEqual(modifiedData.DependentVariables(5:end, :), data.DependentVariables(5:end, :), "Only data within filter should be replaced.");
+            testCase.assertSize(modifiedScience.DependentVariables, size(science.DependentVariables), "Data size should not change.");
+            testCase.verifyTrue(all(ismissing(modifiedScience.XYZ(1:4, :)), "all"), "Data within filter should be replaced.");
+            testCase.verifyEqual(modifiedScience.DependentVariables(5:end, :), science.DependentVariables(5:end, :), "Only data within filter should be replaced.");
         end
 
         % Test that "replace" method replaces data with default filler.
@@ -257,16 +257,49 @@ classdef tScience < matlab.unittest.TestCase
             data = testCase.createTestData();
 
             % Exercise.
-            modifiedData = data.copy();
-            modifiedData.replace(minutes(3), 0);
+            modifiedScience = data.copy();
+            modifiedScience.replace(minutes(3), 0);
 
             % Verify.
-            testCase.assertSize(modifiedData.IndependentVariable, size(data.Time), "Time should not be modified.");
-            testCase.verifyEqual(modifiedData.Time, data.Time, "Time should not be modified.");
+            testCase.assertSize(modifiedScience.IndependentVariable, size(data.Time), "Time should not be modified.");
+            testCase.verifyEqual(modifiedScience.Time, data.Time, "Time should not be modified.");
 
-            testCase.assertSize(modifiedData.DependentVariables, size(data.DependentVariables), "Data size should not change.");
-            testCase.verifyEqual(modifiedData.XYZ(1:4, :), zeros(4, 3), "Data within filter should be replaced.");
-            testCase.verifyEqual(modifiedData.DependentVariables(5:end, :), data.DependentVariables(5:end, :), "Only data within filter should be replaced.");
+            testCase.assertSize(modifiedScience.DependentVariables, size(data.DependentVariables), "Data size should not change.");
+            testCase.verifyEqual(modifiedScience.XYZ(1:4, :), zeros(4, 3), "Data within filter should be replaced.");
+            testCase.verifyEqual(modifiedScience.DependentVariables(5:end, :), data.DependentVariables(5:end, :), "Only data within filter should be replaced.");
+        end
+
+        % Test that PSD can detect sine wave frequency, with default
+        % options.
+        function computePSD_sineWave_default(testCase)
+
+            % Set up.
+            science = testCase.createSineWaveTestData();
+
+            % Execute.
+            psd = science.computePSD();
+
+            % Verify.
+            [~, idxMax] = max([psd.X, psd.Y, psd.Z]);
+
+            testCase.verifyEqual(psd.Frequency(idxMax), [50; 100; 150], "PSD max frequency should match sine wave frequency.");
+        end
+
+        % Test that PSD can detect sine wave frequency, with selected data
+        % only.
+        function computePSD_sineWave_startAndDuration(testCase)
+
+            % Set up.
+            science = testCase.createSineWaveTestData();
+
+            % Execute.
+            psd = science.computePSD(Start = science.Time(10), Duration = milliseconds(500));
+
+            % Verify.
+            [~, idxMax] = max([psd.X, psd.Y, psd.Z]);
+
+            testCase.verifyThat(psd.Frequency(idxMax), matlab.unittest.constraints.IsEqualTo([50; 100; 150], Within = matlab.unittest.constraints.RelativeTolerance(0.1)), ...
+                "PSD max frequency should match sine wave frequency.");
         end
 
         % Test that displaying a single object displays the correct
@@ -274,15 +307,15 @@ classdef tScience < matlab.unittest.TestCase
         function customDisplay_singleObject(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            data.MetaData.DataFrequency = 64;
-            data.MetaData.Mode = "Burst";
-            data.MetaData.Model = "FM4";
-            data.MetaData.Sensor = "FIB";
+            science.MetaData.DataFrequency = 64;
+            science.MetaData.Mode = "Burst";
+            science.MetaData.Model = "FM4";
+            science.MetaData.Sensor = "FIB";
 
             % Exercise.
-            output = evalc("display(data)");
+            output = evalc("display(science)");
 
             % Verify.
             testCase.verifySubstring(eraseTags(output), "FIB (FM4) in Burst (64)", "Science meta data should be included in display.");
@@ -293,15 +326,15 @@ classdef tScience < matlab.unittest.TestCase
         function customDisplay_singleObject_noModel(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            data.MetaData.DataFrequency = 64;
-            data.MetaData.Mode = "Burst";
-            data.MetaData.Model = string.empty();
-            data.MetaData.Sensor = "FIB";
+            science.MetaData.DataFrequency = 64;
+            science.MetaData.Mode = "Burst";
+            science.MetaData.Model = string.empty();
+            science.MetaData.Sensor = "FIB";
 
             % Exercise.
-            output = evalc("display(data)");
+            output = evalc("display(science)");
 
             % Verify.
             testCase.verifySubstring(eraseTags(output), "FIB in Burst (64)", "Science meta data should be included in display.");
@@ -312,15 +345,15 @@ classdef tScience < matlab.unittest.TestCase
         function customDisplay_singleObject_noSensor(testCase)
 
             % Set up.
-            data = testCase.createTestData();
+            science = testCase.createTestData();
 
-            data.MetaData.DataFrequency = 64;
-            data.MetaData.Mode = "Burst";
-            data.MetaData.Model = string.empty();
-            data.MetaData.Sensor = mag.meta.Sensor.empty();
+            science.MetaData.DataFrequency = 64;
+            science.MetaData.Mode = "Burst";
+            science.MetaData.Model = string.empty();
+            science.MetaData.Sensor = mag.meta.Sensor.empty();
 
             % Exercise.
-            output = evalc("display(data)");
+            output = evalc("display(science)");
 
             % Verify.
             testCase.verifySubstring(eraseTags(output), "in Burst (64)", "Science meta data should be included in display.");
@@ -330,36 +363,35 @@ classdef tScience < matlab.unittest.TestCase
         function customDisplay_heterogeneous(testCase)
 
             % Set up.
-            data = testCase.createTestData();
-
-            data = [data, data];
+            science = testCase.createTestData();
+            science = [science, science]; %#ok<NASGU>
 
             % Exercise and verify.
-            display(data);
+            evalc("display(science)");
         end
     end
 
     methods (Access = private)
 
-        function cropAndVerify(testCase, data, timeFilter, expectedTimes, expectedData)
+        function cropAndVerify(testCase, science, timeFilter, expectedTimes, expectedData)
 
             % Exercise.
-            data.crop(timeFilter);
+            science.crop(timeFilter);
 
             % Verify.
-            testCase.assertSize(data.IndependentVariable, size(expectedTimes), "Data should be cropped as expected.");
-            testCase.verifyEqual(data.Time, expectedTimes, "Data should be cropped as expected.");
+            testCase.assertSize(science.IndependentVariable, size(expectedTimes), "Data should be cropped as expected.");
+            testCase.verifyEqual(science.Time, expectedTimes, "Data should be cropped as expected.");
 
-            testCase.assertSize(data.DependentVariables, size(expectedData), "Data should be cropped as expected.");
-            testCase.verifyEqual(data.DependentVariables, expectedData, "Data should be cropped as expected.");
+            testCase.assertSize(science.DependentVariables, size(expectedData), "Data should be cropped as expected.");
+            testCase.verifyEqual(science.DependentVariables, expectedData, "Data should be cropped as expected.");
 
-            testCase.verifyEqual(data.MetaData.Timestamp, data.Time(1), "Meta data timestamp should be updated.");
+            testCase.verifyEqual(science.MetaData.Timestamp, science.Time(1), "Meta data timestamp should be updated.");
         end
     end
 
     methods (Static, Access = private)
 
-        function [data, rawData] = createEmptyTestData()
+        function [science, rawData] = createEmptyTestData()
 
             emptyTime = datetime.empty();
             emptyTime.TimeZone = "UTC";
@@ -372,13 +404,28 @@ classdef tScience < matlab.unittest.TestCase
                 sequence = double.empty(0, 1)));
             rawData = table2timetable(rawData, RowTimes = "Time");
 
-            data = mag.Science(rawData, mag.meta.Science(Timestamp = datetime("now", TimeZone = "UTC")));
+            science = mag.Science(rawData, mag.meta.Science(Timestamp = datetime("now", TimeZone = "UTC")));
         end
 
-        function [data, rawData] = createTestData()
+        function [science, rawData] = createTestData()
 
             rawData = timetable(datetime("now", TimeZone = "UTC") + minutes(1:10)', (1:10)', (11:20)', (21:30)', 3 * ones(10, 1), (1:10)', VariableNames = ["x", "y", "z", "range", "sequence"]);
-            data = mag.Science(rawData, mag.meta.Science(Timestamp = datetime("now", TimeZone = "UTC")));
+            science = mag.Science(rawData, mag.meta.Science(Timestamp = datetime("now", TimeZone = "UTC")));
+        end
+
+        function [science, rawData] = createSineWaveTestData()
+
+            num = 1000;
+
+            timestamp = datetime("now", TimeZone = "UTC") + milliseconds(1:num)';
+
+            t = seconds(timestamp - timestamp(1));
+            x = sin(100 * pi * t);
+            y = sin(200 * pi * t);
+            z = sin(300 * pi * t);
+
+            rawData = timetable(timestamp, x, y, z, 3 * ones(num, 1), (1:num)', VariableNames = ["x", "y", "z", "range", "sequence"]);
+            science = mag.Science(rawData, mag.meta.Science(Timestamp = datetime("now", TimeZone = "UTC")));
         end
     end
 end